django/requirements.txt
django==1.7.0
django-secure==1.0
dj-database-url==0.3.0
django-cors-headers==0.13
django-filter==0.7
django-jsonfield==0.9.13
django-model-utils==2.2.0
djangorestframework==2.4.2
psycopg2==2.5.4
purl==1.0
requests==2.4.1
